+++ master/internal/sproto/task_actor.go
@@ -29,18 +29,6 @@ type (
 		// on the Killed logs.
 		Level *string
 	}
-	// ResourcesID is the ID of some set of resources.
-	ResourcesID string
-
-	// ResourcesType is the type of some set of resources. This should be purely informational.
-	ResourcesType string
-
-	// ResourcesStarted contains the information needed by tasks from container started.
-	ResourcesStarted struct {
-		Addresses []cproto.Address
-		// NativeResourcesID is the native Docker hex container ID of the Determined container.
-		NativeResourcesID string
-	}
 
 	// GetResourcesContainerState requests cproto.Container state for a given clump of resources.
 	// If the resources aren't a container, this request returns a failure.
@@ -101,93 +89,5 @@ func (c ContainerLog) ToTaskLog() model.TaskLog {
 		Level:       c.Level,
 		Timestamp:   ptrs.TimePtr(c.Timestamp.UTC()),
 		Log:         c.Message(),
-	}
-}
-
-// ExitCode is the process exit code of the container.
-type ExitCode int
-
-const (
-	// SuccessExitCode is the 0 zero value exit code.
-	SuccessExitCode = 0
-)
-
-// FromContainerExitCode converts an aproto.ExitCode to an ExitCode. ExitCode's type is subject to
-// change - it may become an enum instead where we interpret the type of exit for consumers.
-func FromContainerExitCode(c *aproto.ExitCode) *ExitCode {
-	if c == nil {
-		return nil
-	}
-	ec := ExitCode(*c)
-	return &ec
-}
-
-// FailureType denotes the type of failure that resulted in the container stopping.
-// Each FailureType must be handled by ./internal/task/allocation.go.
-type FailureType string
-
-const (
-	// ContainerFailed denotes that the container ran but failed with a non-zero exit code.
-	ContainerFailed = FailureType("container failed with non-zero exit code")
-
-	// ContainerAborted denotes the container was canceled before it was started.
-	ContainerAborted = FailureType("container was aborted before it started")
-
-	// TaskAborted denotes that the task was canceled before it was started.
-	TaskAborted = FailureType("task was aborted before the task was started")
-
-	// TaskError denotes that the task failed without an associated exit code.
-	TaskError = FailureType("task failed without an associated exit code")
-
-	// AgentFailed denotes that the agent failed while the container was running.
-	AgentFailed = FailureType("agent failed while the container was running")
-
-	// AgentError denotes that the agent failed to launch the container.
-	AgentError = FailureType("agent failed to launch the container")
-
-	// UnknownError denotes an internal error that did not map to a know failure type.
-	UnknownError
-)
-
-// FromContainerFailureType converts an aproto.FailureType to a FailureType. This mapping is not
-// guaranteed to remain one to one; this conversion may do some level of interpretation.
-func FromContainerFailureType(t aproto.FailureType) FailureType {
-	switch t {
-	case aproto.ContainerFailed:
-		return FailureType(t)
-	case aproto.ContainerAborted:
-		return FailureType(t)
-	case aproto.TaskAborted:
-		return FailureType(t)
-	case aproto.TaskError:
-		return FailureType(t)
-	case aproto.AgentFailed:
-		return FailureType(t)
-	case aproto.AgentError:
-		return FailureType(t)
-	default:
-		return FailureType(t)
-	}
-}
-
-// IsRestartableSystemError checks if the error is caused by the system and
-// shouldn't count against `max_restarts`.
-func IsRestartableSystemError(err error) bool {
-	switch contErr := err.(type) {
-	case ResourcesFailure:
-		switch contErr.FailureType {
-		case ContainerFailed, TaskError:
-			return false
-		// Questionable, could be considered failures, but for now we don't.
-		case AgentError, AgentFailed:
-			return true
-		// Definitely not a failure.
-		case TaskAborted, ContainerAborted:
-			return true
-		default:
-			return false
-		}
-	default:
-		return false
 	}
 }
